Update part1.html
part1.html contains the Web page content without style. 
The result of this stage is a web page which contains a form. The form contains various form elements like textarea, submit button, checkboxes, radio that can be selected and in which the data can be entered. It shows the basic layout of the given web page.